+++ beacon_node/lighthouse_network/tests/rpc_tests.rs
